% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataAnime.R
\docType{data}
\name{dataAnime}
\alias{dataAnime}
\title{Data anime}
\format{Data frame with columns
\describe{
\item{Title}{date in character format}
\item{Type}{time in character format}
\item{Episodes}{username in character format}
\item{Status}{tweet content in character format}
\item{Start airing}{should be logical}
\item{End airing}{number of tweet replied}
\item{Starting season}{number of tweet retweeted}
\item{Broadcast time}{number of tweet favorited}
\item{Licensors}{tweet source}
\item{Studios}{tweet source}
\item{Sources}{tweet source}
\item{Genres}{tweet source}
\item{Rating}{tweet source}
\item{Score}{tweet source}
\item{Scored by}{tweet source}
\item{Members}{tweet source}
\item{Favorites}{tweet source}
\item{Description}{tweet source}
}}
\usage{
dataAnime
}
\description{
Information about anime
}
\examples{
  dataAnime
}
\keyword{datasets}
