% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mldlSurvey.R
\docType{data}
\name{mldlSurvey}
\alias{mldlSurvey}
\title{Kaggle ML & DS Survey}
\format{Large list with 3 elements
\describe{
\item{multipleChoiceResponses}{data frame with 35 varibales and 23860 observations}
\item{freeFormResponses}{data frame with 395 varibales and 23860 observations}
\item{SurveySchema}{data frame with 52 variables and 52 observations}
}}
\source{
https://www.kaggle.com/kaggle/kaggle-survey-2018/downloads/kaggle-survey-2018.zip/5
}
\usage{
mldlSurvey
}
\description{
Kaggle ML & DS Survey 2018
}
\examples{
  mldlSurvey
}
\keyword{datasets}
