% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normaliselexicon.R
\docType{data}
\name{normaliselexicon}
\alias{normaliselexicon}
\title{Normalisation lexicon}
\format{Data frame with columns
\describe{
\item{earn}{double}
\item{height}{double in inch}
}}
\usage{
normaliselexicon
}
\description{
Normalisation lexicon in bahasa Indonesia
}
\examples{
  normaliselexicon
}
\keyword{datasets}
