% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pendapatan.R
\docType{data}
\name{pendapatan}
\alias{pendapatan}
\title{Pendapatan per gender}
\format{Data frame with columns
\describe{
\item{earn}{double}
\item{height}{double in inch}
\item{sex}{character}
\item{ed}{character}
\item{age}{double}
\item{race}{character}
}}
\usage{
pendapatan
}
\description{
Pendapatan berdasarkan gender
}
\examples{
  pendapatan
}
\keyword{datasets}
