% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords_id.R
\docType{data}
\name{stopwords_id}
\alias{stopwords_id}
\title{Stopwords bahasa Indonesia}
\format{Data frame with columns
\describe{
\item{V1}{Term in character format}
}}
\usage{
stopwords_id
}
\description{
One column stopwords in Bahasa Indonesia
}
\examples{
  stopwords_id
}
\keyword{datasets}
