% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataAnime.R
\docType{data}
\name{dataAnime}
\alias{dataAnime}
\title{Data anime}
\format{Data frame with columns
\describe{
\item{Title}{chr}
\item{Type}{chr}
\item{Episodes}{chr}
\item{Status}{chr}
\item{Start airing}{chr}
\item{End airing}{chr}
\item{Starting season}{chr}
\item{Broadcast time}{chr}
\item{Licensors}{chr}
\item{Studios}{chr}
\item{Sources}{chr}
\item{Genres}{chr}
\item{Rating}{chacrater}
\item{Score}{number}
\item{Scored by}{number}
\item{Members}{number}
\item{Favorites}{number}
\item{Description}{character description}
}}
\source{
https://www.kaggle.com/canggih/anime-data-score-staff-synopsis-and-genre
}
\usage{
dataAnime
}
\description{
Information about anime
}
\examples{
  dataAnime
}
\keyword{datasets}
