% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kedataTwit.R
\docType{data}
\name{kedataTwit}
\alias{kedataTwit}
\title{Tweet sample}
\format{Data frame with columns
\describe{
\item{created_at}{date in character format}
\item{time}{time in character format}
\item{username}{username in character format}
\item{text}{tweet content in character format}
\item{is_retweet}{should be logical}
\item{reply}{number of tweet replied}
\item{retweet}{number of tweet retweeted}
\item{favorite}{number of tweet favorited}
\item{source}{tweet source}
}}
\source{
https://www.transtats.bts.gov/DL_SelectFields.asp?Table_ID=236
}
\usage{
kedataTwit
}
\description{
Look up airline names from their carrier codes.
}
\examples{
  kedataTwit
}
\keyword{datasets}
